requirements.txt
pandas
matplotlib
requests
dash
scikit-learn
flask-cors
mysql-connector-python
dotenv
google-generativeai
statsmodels
tensorflow